Adjust anamcore.

+++ AnamcoreManager.cs
@@ -71,7 +71,7 @@ namespace AnamCore
                 Plugin.PluginLog.Error(e, e.Message);
             }
         }
-        public async void TriggerEmote(nint character, ushort animationId)
+        public async void TriggerEmote(nint character, uint animationId)
         {
             try
             {
@@ -87,7 +87,7 @@ namespace AnamCore
             }
         }
 
-        public async void TriggerEmoteTimed(ICharacter character, ushort animationId, int time = 2000)
+        public async void TriggerEmoteTimed(ICharacter character, uint animationId, int time = 2000)
         {
             try
             {
@@ -96,7 +96,7 @@ namespace AnamCore
                 var animationMemory = actorMemory.Animation;
                 if (animationMemory.BaseOverride != animationId)
                 {
-                    animationMemory!.BaseOverride = animationId;
+                    animationMemory!.BaseOverride = (ushort)animationId;
                     MemoryService.Write(animationMemory.GetAddressOfProperty(nameof(AnimationMemory.BaseOverride)), animationId, "Base Override");
                 }
                 byte originalMode = MemoryService.Read<byte>(actorMemory.GetAddressOfProperty(nameof(ActorMemory.CharacterModeRaw)));
